Add files via upload
Need to improve the final selection, as of now it just spits out two matrices for the front and rear breaking torques (rows are pedal ratio and the columns are the bore for the master cylinder). The user compares each matrix to its corresponding required braking torque * 1.2 (increase by 20%  for error margin due to large estimates of vehicle weight and CG) to find the best pedal ratio and master cylinders for the front and rear braking systems.